feat(OY2-25568): RHF Doc typing
