+++ src/services/ui/src/pages/form/proto.tsx
@@ -1,4 +1,4 @@
-import { Document } from "@/components/RHF/RHFInput";
+import { Document } from "@/components/RHF/types";
 
 export const ABP1: Document = {
   header: "ABP1: Alternative Benefit Plan",
@@ -14,8 +14,9 @@ export const ABP1: Document = {
               rhf: "Input",
               name: "alt_benefit_plan_population_name",
               label: "Alternative Benefit Plan population name",
-              placeholder: "enter name",
-
+              props: {
+                placeholder: "enter name",
+              },
               dependency: {
                 // example of a value changing field, with multi-conditions
                 conditions: [
@@ -45,40 +46,44 @@ export const ABP1: Document = {
               fields: [
                 {
                   rhf: "Select",
-                  placeholder: "enter name",
-                  className: "w-[300px]",
-                  label: "Eligibility group",
                   name: "eligibility_group",
-                  options: [
-                    {
-                      label: "Extended medicaid due to earnings",
-                      value: "option1",
-                    },
-                    {
-                      label: "Extended medicaid due to bling",
-                      value: "option2",
-                    },
-                    {
-                      label: "Extended medicaid due to bankrupcy",
-                      value: "option3",
-                    },
-                  ],
+                  label: "Eligibility group",
+                  props: {
+                    placeholder: "enter name",
+                    className: "w-[300px]",
+                    options: [
+                      {
+                        label: "Extended medicaid due to earnings",
+                        value: "option1",
+                      },
+                      {
+                        label: "Extended medicaid due to bling",
+                        value: "option2",
+                      },
+                      {
+                        label: "Extended medicaid due to bankrupcy",
+                        value: "option3",
+                      },
+                    ],
+                  },
                 },
                 {
                   rhf: "Select",
                   name: "mandatory_voluntary",
-                  className: "w-[200px]",
                   label: "Mandatory or voluntary",
-                  options: [
-                    {
-                      label: "Mandatory",
-                      value: "mandatory",
-                    },
-                    {
-                      label: "Voluntary",
-                      value: "voluntary",
-                    },
-                  ],
+                  props: {
+                    className: "w-[200px]",
+                    options: [
+                      {
+                        label: "Mandatory",
+                        value: "mandatory",
+                      },
+                      {
+                        label: "Voluntary",
+                        value: "voluntary",
+                      },
+                    ],
+                  },
                 },
               ],
             },
@@ -100,14 +105,15 @@ export const ABP1: Document = {
           slots: [
             {
               rhf: "Select",
-              label: "Alternative Benefit Plan population name",
               name: "is_enrollment_available",
-              className: "w-[150px]",
-              placeholder: "Select",
-              options: [
-                { label: "Yes", value: "yes" },
-                { label: "No", value: "no" },
-              ],
+              label: "Alternative Benefit Plan population name",
+              props: {
+                className: "w-[150px]",
+                options: [
+                  { label: "Yes", value: "yes" },
+                  { label: "No", value: "no" },
+                ],
+              },
             },
           ],
         },
@@ -134,180 +140,198 @@ export const ABP1: Document = {
               rhf: "Checkbox",
               name: "target_criteria",
               label: "Mandatory or voluntary",
-              options: [
-                {
-                  value: "income_standard",
-                  label: "Income standard",
-                  form: [
-                    {
-                      description: "Income standard target",
-                      slots: [
-                        {
-                          rhf: "Radio",
-                          name: "income_target",
-                          options: [
-                            {
-                              label:
-                                "Households with income at or below the standard",
-                              value: "income_target_below",
-                            },
-                            {
-                              label:
-                                "Households with income above the standard",
-                              value: "income_target_above",
-                            },
-                          ],
-                        },
-                      ],
-                    },
-                    {
-                      description: "Income standard definition",
-                      slots: [
-                        {
-                          rhf: "Radio",
-                          name: "income_definition",
-                          options: [
-                            {
-                              label: "A percentage",
-                              value: "income_definition_percentage",
-                            },
-                            {
-                              label: "A specific amount",
-                              value: "income_definition_specific",
-                              slots: [
+              props: {
+                options: [
+                  {
+                    value: "income_standard",
+                    label: "Income standard",
+                    form: [
+                      {
+                        description: "Income standard target",
+                        slots: [
+                          {
+                            rhf: "Radio",
+                            name: "income_target",
+                            props: {
+                              options: [
                                 {
-                                  rhf: "Radio",
-                                  name: "income_definition_specific",
-                                  options: [
+                                  label:
+                                    "Households with income at or below the standard",
+                                  value: "income_target_below",
+                                },
+                                {
+                                  label:
+                                    "Households with income above the standard",
+                                  value: "income_target_above",
+                                },
+                              ],
+                            },
+                          },
+                        ],
+                      },
+                      {
+                        description: "Income standard definition",
+                        slots: [
+                          {
+                            rhf: "Radio",
+                            name: "income_definition",
+                            props: {
+                              options: [
+                                {
+                                  label: "A percentage",
+                                  value: "income_definition_percentage",
+                                },
+                                {
+                                  label: "A specific amount",
+                                  value: "income_definition_specific",
+                                  slots: [
                                     {
-                                      label: "Statewide standard",
-                                      value: "statewide_standard",
-                                      form: [
-                                        {
-                                          slots: [
-                                            {
-                                              rhf: "FieldArray",
-                                              name: "income_definition_specific_statewide",
-                                              fields: [
-                                                {
-                                                  rhf: "Input",
-                                                  placeholder: "enter size",
-                                                  className: "w-[300px]",
-                                                  label: "Household Size",
-                                                  name: "household_size",
-                                                },
-                                                {
-                                                  rhf: "Input",
-                                                  name: "standard",
-                                                  className: "w-[200px]",
-                                                  placenholder: "enter amount",
-                                                  label: "Standard ($)",
-                                                },
-                                              ],
-                                            },
-                                          ],
-                                        },
-                                        {
-                                          description:
-                                            "Is there an additional incremental amount",
-                                          slots: [
-                                            {
-                                              rhf: "Switch",
-                                              name: "is_incremental_amount",
-                                            },
-                                          ],
-                                        },
-                                        {
-                                          description:
-                                            "Enter incremental dollar amount",
-                                          slots: [
-                                            {
-                                              rhf: "Input",
-                                              name: "doller_incremental_amount",
-                                            },
-                                          ],
-                                        },
-                                      ],
-                                    },
-                                    {
-                                      label: "Standard Varies by region",
-                                      value: "region_standard",
-                                    },
-                                    {
-                                      label:
-                                        "standard varies by living arrangement",
-                                      value: "living_standard",
-                                    },
-                                    {
-                                      label:
-                                        "standard varies by some other way",
-                                      value: "other_standard",
+                                      rhf: "Radio",
+                                      name: "income_definition_specific",
+                                      props: {
+                                        options: [
+                                          {
+                                            label: "Statewide standard",
+                                            value: "statewide_standard",
+                                            form: [
+                                              {
+                                                slots: [
+                                                  {
+                                                    rhf: "FieldArray",
+                                                    name: "income_definition_specific_statewide",
+                                                    fields: [
+                                                      {
+                                                        rhf: "Input",
+                                                        label: "Household Size",
+                                                        name: "household_size",
+                                                        props: {
+                                                          placeholder:
+                                                            "enter size",
+                                                          className:
+                                                            "w-[300px]",
+                                                        },
+                                                      },
+                                                      {
+                                                        rhf: "Input",
+                                                        name: "standard",
+                                                        label: "Standard ($)",
+                                                        props: {
+                                                          className:
+                                                            "w-[200px]",
+                                                          placenholder:
+                                                            "enter amount",
+                                                        },
+                                                      },
+                                                    ],
+                                                  },
+                                                ],
+                                              },
+                                              {
+                                                description:
+                                                  "Is there an additional incremental amount",
+                                                slots: [
+                                                  {
+                                                    rhf: "Switch",
+                                                    name: "is_incremental_amount",
+                                                  },
+                                                ],
+                                              },
+                                              {
+                                                description:
+                                                  "Enter incremental dollar amount",
+                                                slots: [
+                                                  {
+                                                    rhf: "Input",
+                                                    name: "doller_incremental_amount",
+                                                  },
+                                                ],
+                                              },
+                                            ],
+                                          },
+                                          {
+                                            label: "Standard Varies by region",
+                                            value: "region_standard",
+                                          },
+                                          {
+                                            label:
+                                              "standard varies by living arrangement",
+                                            value: "living_standard",
+                                          },
+                                          {
+                                            label:
+                                              "standard varies by some other way",
+                                            value: "other_standard",
+                                          },
+                                        ],
+                                      },
                                     },
                                   ],
                                 },
                               ],
                             },
-                          ],
-                        },
-                      ],
-                    },
-                  ],
-                },
-                {
-                  value: "health",
-                  label: "Disease, condition, diagnosis, or disorder",
-                  slots: [
-                    {
-                      rhf: "Checkbox",
-                      name: "health_conditions",
-                      options: [
-                        {
-                          label: "Physical disability",
-                          value: "physical_disability",
-                        },
-                        {
-                          label: "Brain Injury",
-                          value: "brain_injury",
-                        },
-                        {
-                          label: "HIV / AIDS",
-                          value: "hiv_aids",
-                        },
-                        {
-                          label: "Medically frail",
-                          value: "ben",
-                        },
-                        {
-                          label: "Technology dependent",
-                          value: "technology_dependent",
-                        },
-                        {
-                          label: "Other",
-                          value: "other",
-                          slots: [
+                          },
+                        ],
+                      },
+                    ],
+                  },
+                  {
+                    value: "health",
+                    label: "Disease, condition, diagnosis, or disorder",
+                    slots: [
+                      {
+                        rhf: "Checkbox",
+                        name: "health_conditions",
+                        props: {
+                          options: [
                             {
-                              rhf: "Input",
-                              name: "other_description",
-                              label: "Describe",
+                              label: "Physical disability",
+                              value: "physical_disability",
+                            },
+                            {
+                              label: "Brain Injury",
+                              value: "brain_injury",
+                            },
+                            {
+                              label: "HIV / AIDS",
+                              value: "hiv_aids",
+                            },
+                            {
+                              label: "Medically frail",
+                              value: "ben",
+                            },
+                            {
+                              label: "Technology dependent",
+                              value: "technology_dependent",
+                            },
+                            {
+                              label: "Other",
+                              value: "other",
+                              slots: [
+                                {
+                                  rhf: "Input",
+                                  name: "other_description",
+                                  label: "Describe",
+                                },
+                              ],
                             },
                           ],
                         },
-                      ],
-                    },
-                  ],
-                },
-                {
-                  label: "Other targeting criteria",
-                  value: "other_targeting_criteria",
-                  slots: [
-                    {
-                      rhf: "Input",
-                      name: "other_targeting_criteria_description",
-                      label: "Describe",
-                    },
-                  ],
-                },
-              ],
+                      },
+                    ],
+                  },
+                  {
+                    label: "Other targeting criteria",
+                    value: "other_targeting_criteria",
+                    slots: [
+                      {
+                        rhf: "Input",
+                        name: "other_targeting_criteria_description",
+                        label: "Describe",
+                      },
+                    ],
+                  },
+                ],
+              },
             },
           ],
         },
@@ -322,12 +346,14 @@ export const ABP1: Document = {
           slots: [
             {
               rhf: "Select",
-              className: "w-[150px]",
               name: "is_geographic_area",
-              options: [
-                { label: "Yes", value: "yes" },
-                { label: "No", value: "no" },
-              ],
+              props: {
+                className: "w-[150px]",
+                options: [
+                  { label: "Yes", value: "yes" },
+                  { label: "No", value: "no" },
+                ],
+              },
             },
           ],
         },
@@ -337,12 +363,14 @@ export const ABP1: Document = {
             {
               rhf: "Radio",
               name: "geographic_variation",
-              options: [
-                { label: "By country", value: "by_country" },
-                { label: "By region", value: "by_region" },
-                { label: "By city or town", value: "by_city_town" },
-                { label: "Other geographic area", value: "other" },
-              ],
+              props: {
+                options: [
+                  { label: "By country", value: "by_country" },
+                  { label: "By region", value: "by_region" },
+                  { label: "By city or town", value: "by_city_town" },
+                  { label: "Other geographic area", value: "other" },
+                ],
+              },
             },
           ],
         },
@@ -403,20 +431,26 @@ export const ABP1: Document = {
             {
               name: "example2_1",
               label: "Example 2.1",
-              className: "w-80",
               rhf: "Input",
+              props: {
+                className: "w-80",
+              },
             },
             {
               name: "example2_2",
               label: "Example 2.2",
-              className: "w-30",
               rhf: "Input",
+              props: {
+                className: "w-30",
+              },
             },
             {
               name: "example2_3",
               label: "Example 2.3",
-              className: "w-120",
               rhf: "Input",
+              props: {
+                className: "w-120",
+              },
             },
           ],
         },
